pyproject.toml
[project]
name = "medical-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohappyeyeballs==2.6.1 ; python_full_version >= '3.9'",
    "aiohttp==3.12.14 ; python_full_version >= '3.9'",
    "aiosignal==1.4.0 ; python_full_version >= '3.9'",
    "altair==5.5.0 ; python_full_version >= '3.9'",
    "annotated-types==0.7.0 ; python_full_version >= '3.8'",
    "anyio==4.9.0 ; python_full_version >= '3.9'",
    "attrs==25.3.0 ; python_full_version >= '3.8'",
    "blinker==1.9.0 ; python_full_version >= '3.9'",
    "cachetools==6.1.0 ; python_full_version >= '3.9'",
    "certifi==2025.7.9 ; python_full_version >= '3.7'",
    "charset-normalizer==3.4.2 ; python_full_version >= '3.7'",
    "click==8.2.1 ; python_full_version >= '3.10'",
    "dataclasses-json==0.6.7 ; python_full_version >= '3.7' and python_full_version < '4'",
    "distro==1.9.0 ; python_full_version >= '3.6'",
    "faiss-cpu==1.11.0 ; python_full_version >= '3.9'",
    "filelock==3.18.0 ; python_full_version >= '3.9'",
    "frozenlist==1.7.0 ; python_full_version >= '3.9'",
    "fsspec==2025.5.1 ; python_full_version >= '3.9'",
    "gitdb==4.0.12 ; python_full_version >= '3.7'",
    "gitpython==3.1.44 ; python_full_version >= '3.7'",
    "groq==0.29.0 ; python_full_version >= '3.8'",
    "h11==0.16.0 ; python_full_version >= '3.8'",
    "hf-xet==1.1.5 ; platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'arm64' or platform_machine == 'x86_64'",
    "httpcore==1.0.9 ; python_full_version >= '3.8'",
    "httpx==0.28.1 ; python_full_version >= '3.8'",
    "httpx-sse==0.4.1 ; python_full_version >= '3.9'",
    "huggingface-hub==0.33.2 ; python_full_version >= '3.8'",
    "idna==3.10 ; python_full_version >= '3.6'",
    "jinja2==3.1.6 ; python_full_version >= '3.7'",
    "joblib==1.5.1 ; python_full_version >= '3.9'",
    "jsonpatch==1.33 ; python_full_version >= '2.7'",
    "jsonpointer==3.0.0 ; python_full_version >= '3.7'",
    "jsonschema==4.24.0 ; python_full_version >= '3.9'",
    "jsonschema-specifications==2025.4.1 ; python_full_version >= '3.9'",
    "langchain==0.3.26 ; python_full_version >= '3.9'",
    "langchain-community==0.3.27 ; python_full_version >= '3.9'",
    "langchain-core==0.3.68 ; python_full_version >= '3.9'",
    "langchain-groq==0.3.5 ; python_full_version >= '3.9'",
    "langchain-huggingface==0.3.0 ; python_full_version >= '3.9'",
    "langchain-text-splitters==0.3.8 ; python_full_version >= '3.9' and python_full_version < '4'",
    "langsmith==0.4.4 ; python_full_version >= '3.9'",
    "markupsafe==3.0.2 ; python_full_version >= '3.9'",
    "marshmallow==3.26.1 ; python_full_version >= '3.9'",
    "mpmath==1.3.0",
    "multidict==6.6.3 ; python_full_version >= '3.9'",
    "mypy-extensions==1.1.0 ; python_full_version >= '3.8'",
    "narwhals==1.46.0 ; python_full_version >= '3.9'",
    "networkx==3.5 ; python_full_version >= '3.11'",
    "numpy==2.3.1 ; python_full_version < '3.13'",
    "orjson==3.10.18 ; platform_python_implementation != 'PyPy'",
    "packaging==24.2 ; python_full_version >= '3.8'",
    "pandas==2.3.1 ; python_full_version >= '3.9'",
    "pillow==11.3.0 ; python_full_version >= '3.9'",
    "propcache==0.3.2 ; python_full_version >= '3.9'",
    "protobuf==6.31.1 ; python_full_version >= '3.9'",
    "pyarrow==20.0.0 ; python_full_version >= '3.9'",
    "pydantic==2.11.7 ; python_full_version >= '3.9'",
    "pydantic-core==2.33.2 ; python_full_version >= '3.9'",
    "pydantic-settings==2.10.1 ; python_full_version >= '3.9'",
    "pydeck==0.9.1 ; python_full_version >= '3.8'",
    "pypdf==5.7.0 ; python_full_version >= '3.8'",
    "python-dateutil==2.9.0.post0 ; python_full_version >= '2.7'",
    "python-dotenv==1.1.1 ; python_full_version >= '3.9'",
    "pytz==2025.2",
    "pyyaml==6.0.2 ; python_full_version >= '3.8'",
    "referencing==0.36.2 ; python_full_version >= '3.9'",
    "regex==2024.11.6 ; python_full_version >= '3.8'",
    "requests==2.32.4 ; python_full_version >= '3.8'",
    "requests-toolbelt==1.0.0 ; python_full_version >= '2.7'",
    "rpds-py==0.26.0 ; python_full_version >= '3.9'",
    "safetensors==0.5.3 ; python_full_version >= '3.7'",
    "scikit-learn==1.7.0 ; python_full_version >= '3.10'",
    "scipy==1.16.0 ; python_full_version >= '3.11'",
    "sentence-transformers==5.0.0 ; python_full_version >= '3.9'",
    "setuptools==80.9.0 ; python_full_version >= '3.12'",
    "six==1.17.0 ; python_full_version >= '2.7'",
    "smmap==5.0.2 ; python_full_version >= '3.7'",
    "sniffio==1.3.1 ; python_full_version >= '3.7'",
    "sqlalchemy==2.0.41 ; python_full_version >= '3.7'",
    "streamlit==1.46.1 ; python_full_version > '3.9.7' or (python_full_version >= '3.9' and python_full_version < '3.9.7')",
    "sympy==1.14.0 ; python_full_version >= '3.9'",
    "tenacity==9.1.2 ; python_full_version >= '3.9'",
    "threadpoolctl==3.6.0 ; python_full_version >= '3.9'",
    "tokenizers==0.21.2 ; python_full_version >= '3.9'",
    "toml==0.10.2 ; python_full_version >= '2.6'",
    "torch==2.7.1 ; python_full_version >= '3.9'",
    "tornado==6.5.1 ; python_full_version >= '3.9'",
    "tqdm==4.67.1 ; python_full_version >= '3.7'",
    "transformers==4.53.1 ; python_full_version >= '3.9'",
    "typing-extensions==4.14.1 ; python_full_version >= '3.9'",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.1 ; python_full_version >= '3.9'",
    "tzdata==2025.2 ; python_full_version >= '2'",
    "urllib3==2.5.0 ; python_full_version >= '3.9'",
    "yarl==1.20.1 ; python_full_version >= '3.9'",
    "zstandard==0.23.0 ; python_full_version >= '3.8'",
]
